tools/documentation:  Move content to docs-gen
Due to the split of vss-tools (previous tools/ dir) into a separate
repository, we want to move everything *not* tool-related away from that
repository.  The markdown source files and related layout and template
for this documentation will be stored in this main repository instead.

NOTE: history of this documentation is unfortunately lost because this
move is not recognized as a move since it was now done after the subtree
split into two unrelated repos.

Please refer to the tools/documentation directory, starting from
f31feebdb4 and earlier to study the history or to git-blame those files.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>